Final pre-beta commit: all procedures in provision.py have been changed. Manual site deploy via zip file now semi-automatic

+++ pages/action.py
@@ -7,22 +7,30 @@ action_bp = Blueprint("action", __name__)
 @action_bp.route("/action", methods=['POST'])
 @login_required
 def do_action():
-    if (request.form.get('delete')):
+    #sites delete block
+    if (request.form.get('delete') and not request.form.get('selected')):
         delete_site(request.form['delete'].strip())
+    elif (request.form.get('delete') and request.form.get('selected')):
+        array = request.form.getlist("selected")
+        del_selected_sites(request.form['delete'].strip(),array)
+        return redirect(f"/",301)
+    #sites actions
     elif (request.form.get('disable')):
         disable_site(request.form['disable'].strip())
     elif (request.form.get('enable')):
         enable_site(request.form['enable'].strip())
-    elif (request.form.get('selected')):
+    #redirects management block
+    elif (request.form.get('del_redir') and not request.form.get('selected')):
+        del_redirect(request.form['del_redir'].strip(),request.form['sitename'].strip())
+        return redirect(f"/redirects_manager?site={request.form['sitename'].strip()}",301)
+    elif (request.form.get('del_redir') and request.form.get('selected')):
         array = request.form.getlist("selected")
         del_selected_redirects(array,request.form['sitename'].strip())
-        return redirect(f"/redirects_manager?site={request.form['sitename'].strip()}",301)
-    elif (request.form.get('del_redir')):
-        del_redirect(request.form['del_redir'].strip(),request.form['sitename'].strip())
         return redirect(f"/redirects_manager?site={request.form['sitename'].strip()}",301)
     elif (request.form.get('applyChanges')):
         applyChanges(request.form['sitename'].strip())
         return redirect(f"/redirects_manager?site={request.form['sitename'].strip()}",301)
+    #Git block
     elif (request.form.get('gitPull') and not request.form.get('selected')):
         makePull(request.form['gitPull'].strip())
     elif (request.form.get('gitPull') and request.form.get('selected')):
